full-turn-shift/applications/requirements.txt
faster_whisper
pyaudio
torch
pydub
matplotlib
soundfile==0.12.1
torchaudio
webrtcvad
librosa
google-cloud-speech
pydub